+++ internal/models/user.go
@@ -106,7 +106,7 @@ type UpdateProfileInfo struct {
 
 type Donate struct {
 	CreatorID  uuid.UUID `json:"creator_id"`
-	MoneyCount int       `json:"money_count"`
+	MoneyCount int64     `json:"money_count"`
 }
 
 func (creatorInfo *BecameCreatorInfo) IsValid() bool {
